+++ CHANGELOG.md
@@ -351,6 +351,10 @@
 - Fix bug that prevented delivery partners from submitting a report.
 
 ## [unreleased]
+- Collect financial quarter and year for planned disbursements (forecasts)
+  instead of start and end dates (the values of which are calculated).
+- No longer collect start and end dates for planned disbursements (forecasts)
+  through the application interface.
 
 [unreleased]: https://github.com/UKGovernmentBEIS/beis-report-official-development-assistance/compare/release-20...HEAD
 [release-20]: https://github.com/UKGovernmentBEIS/beis-report-official-development-assistance/compare/release-19...release-20